+++ python/GafferUI/GLWidget.py
@@ -66,6 +66,7 @@ from Qt import QtCore
 # if we can coax the project into bridging Qt 5/6 instead of 4/5?
 from PySide2 import QtGui
 from PySide2 import QtWidgets
+from Qt import QtOpenGL
 
 ## The GLWidget is a base class for all widgets which wish to draw using OpenGL.
 # Derived classes override the _draw() method to achieve this.
@@ -84,20 +85,20 @@ class GLWidget( GafferUI.Widget ) :
 	# even when not requested.
 	def __init__( self, bufferOptions = set(), **kw ) :
 
-		format = QtGui.QSurfaceFormat()
+		_format = QtGui.QSurfaceFormat()
 
 		if self.BufferOptions.Alpha in bufferOptions :
-			format.setAlphaBufferSize( 8 )
+			_format.setAlphaBufferSize( 8 )
 		if self.BufferOptions.Depth in bufferOptions :
-			format.setDepthBufferSize( 24 )
+			_format.setDepthBufferSize( 24 )
 
 		self.__multisample = self.BufferOptions.AntiAlias in bufferOptions
 		if self.__multisample:
-			format.setSamples( 8 )
+			_format.setSamples( 8 )
 
-		format.setVersion( 2, 1 )
+		_format.setVersion( 2, 1 )
 
-		graphicsView = _GLGraphicsView( format )
+		graphicsView = _GLGraphicsView( _format )
 		self.__graphicsScene = _GLGraphicsScene( graphicsView, Gaffer.WeakMethod( self.__draw ) )
 		graphicsView.setScene( self.__graphicsScene )
 
@@ -201,18 +202,14 @@ class GLWidget( GafferUI.Widget ) :
 
 class _GLGraphicsView( QtWidgets.QGraphicsView ) :
 
-	def __init__( self, format ) :
+	def __init__( self, _format ) :
 
 		QtWidgets.QGraphicsView.__init__( self )
 
 		self.setHorizontalScrollBarPolicy( QtCore.Qt.ScrollBarAlwaysOff )
 		self.setVerticalScrollBarPolicy( QtCore.Qt.ScrollBarAlwaysOff )
 
-		glWidget = QtWidgets.QOpenGLWidget()
-		# Avoid `QOpenGLFramebufferObject: Framebuffer incomplete attachment`
-		# errors caused by Qt trying to make a framebuffer with zero size.
-		glWidget.setMinimumSize( 1, 1 )
-		glWidget.setFormat( format )
+		glWidget = self.__createGLWidget( _format )
 		self.setViewport( glWidget )
 		self.setViewportUpdateMode( self.FullViewportUpdate )
 
@@ -248,6 +245,90 @@ class _GLGraphicsView( QtWidgets.QGraphicsView ) :
 				return
 
 		QtWidgets.QFrame.keyPressEvent( self, event )
+
+	@classmethod
+	def __createGLWidget( cls, _format ) :
+
+		# try to make a host specific widget if necessary.
+		result = cls.__createMayaQGLWidget( _format )
+		if result is not None :
+			return result
+
+		result = cls.__createHoudiniQGLWidget( _format )
+		if result is not None :
+			return result
+
+		glWidget = QtWidgets.QOpenGLWidget()
+		# Avoid `QOpenGLFramebufferObject: Framebuffer incomplete attachment`
+		# errors caused by Qt trying to make a framebuffer with zero size.
+		glWidget.setMinimumSize( 1, 1 )
+		glWidget.setFormat( _format )
+
+		return glWidget
+
+	@classmethod
+	def __createQGLFormat( cls, _format ):
+		"""
+		Create a QGLFormat based on the configuration of
+		QSufraceFormat where possible.
+		"""
+		qGLFormat = QtOpenGL.QGLFormat()
+		# note that here we are not using _format.hasAlpha()
+		# because it always return False ( probably uninitalized ).
+		qGLFormat.setRgba( True )
+
+		if _format.samples() > 1:
+			qGLFormat.setSampleBuffers( True )
+
+		return qGLFormat
+
+	@classmethod
+	def __createHostedQGLWidget( cls, _format ) :
+
+		# When running Gaffer embedded in a host application such as Maya
+		# or Houdini, we want to be able to share OpenGL resources between
+		# gaffer uis and host viewport uis, because IECoreGL will be used
+		# in both. So we implement our own QGLContext class which creates a
+		# context which shares with the host. The custom QGLContext is
+		# implemented in GLWidgetBinding.cpp, and automatically shares with
+		# the context which is current at the time of its creation. The host
+		# context should therefore be made current before calling this
+		# method.
+
+		qGLFormat = cls.__createQGLFormat( _format )
+		result = QtOpenGL.QGLWidget()
+		_GafferUI._glWidgetSetHostedContext( GafferUI._qtAddress( result ), GafferUI._qtAddress( qGLFormat ) )
+		return result
+
+	@classmethod
+	def __createMayaQGLWidget( cls, _format ) :
+
+		try :
+			import maya.OpenMayaRender
+		except ImportError :
+			# we're not in maya - createGLWidget() will just make a
+			# normal widget.
+			return None
+
+		mayaRenderer = maya.OpenMayaRender.MHardwareRenderer.theRenderer()
+		mayaRenderer.makeResourceContextCurrent( mayaRenderer.backEndString() )
+		return cls.__createHostedQGLWidget( _format )
+
+	@classmethod
+	def __createHoudiniQGLWidget( cls, _format ) :
+
+		try :
+			import hou
+		except ImportError :
+			# we're not in houdini - createGLWidget() will just make a
+			# normal widget.
+			return None
+
+		import IECoreHoudini
+
+		# Force the Houdini GL context to be current, and share it.
+		IECoreHoudini.makeMainGLContextCurrent()
+		return cls.__createHostedQGLWidget( _format )
 
 class _GLGraphicsScene( QtWidgets.QGraphicsScene ) :